[add]ローディングのテストを追加
